requirements.txt
pandas
chromadb
db-dtypes
streamlit
sentence-transformers
google-cloud-aiplatform
google-cloud-bigquery[pandas]